chore: move string func from base/io to base/string_func
