+++ src/common/base/src/base/mod.rs
@@ -52,10 +52,13 @@ pub use shutdown_signal::SignalType;
 pub use singleton_instance::GlobalInstance;
 pub use stop_handle::StopHandle;
 pub use stoppable::Stoppable;
+pub use string_func::convert_byte_size;
+pub use string_func::convert_number_size;
 pub use string_func::escape_for_key;
 pub use string_func::mask_string;
 pub use string_func::replace_nth_char;
 pub use string_func::unescape_for_key;
+pub use string_func::unescape_string;
 pub use thread::Thread;
 pub use thread::ThreadJoinHandle;
 pub use thread_pool::TaskJoinHandler;
