+++ src/common/io/src/lib.rs
@@ -35,5 +35,5 @@ mod file_split;
 mod format_settings;
 mod options_deserializer;
 mod position;
+mod serialization;
 mod stat_buffer;
-mod utils;
