+++ src/common/io/src/prelude.rs
@@ -31,5 +31,5 @@ pub use crate::format_settings::FormatSettings;
 pub use crate::options_deserializer::OptionsDeserializer;
 pub use crate::options_deserializer::OptionsDeserializerError;
 pub use crate::position::*;
+pub use crate::serialization::*;
 pub use crate::stat_buffer::StatBuffer;
-pub use crate::utils::*;
