+++ src/common/io/tests/it/main.rs
@@ -18,4 +18,3 @@ mod binary_write;
 mod buffer;
 mod cursor_ext;
 mod options_deserializer;
-mod utils;
