+++ src/query/service/src/servers/http/v1/load.rs
@@ -18,11 +18,11 @@ use std::sync::Arc;
 use common_base::base::tokio;
 use common_base::base::tokio::io::AsyncRead;
 use common_base::base::tokio::io::AsyncReadExt;
+use common_base::base::unescape_string;
 use common_base::base::ProgressValues;
 use common_base::base::TrySpawn;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_io::prelude::unescape_string;
 use common_pipeline_sources::processors::sources::input_formats::InputContext;
 use common_pipeline_sources::processors::sources::input_formats::StreamingReadBatch;
 use futures::StreamExt;
