+++ src/query/service/src/servers/mysql/mysql_interactive_worker.rs
@@ -16,6 +16,8 @@ use std::marker::PhantomData;
 use std::sync::Arc;
 use std::time::Instant;
 
+use common_base::base::convert_byte_size;
+use common_base::base::convert_number_size;
 use common_base::base::tokio::io::AsyncWrite;
 use common_base::base::TrySpawn;
 use common_config::DATABEND_COMMIT_VERSION;
@@ -24,7 +26,6 @@ use common_datablocks::SendableDataBlockStream;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_exception::ToErrorCode;
-use common_io::prelude::*;
 use common_users::CertifiedInfo;
 use common_users::UserApiProvider;
 use futures_util::StreamExt;
